requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
sentence-transformers==3.2.1
scikit-learn==1.5.2
numpy==1.26.4
groq==0.11.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
rich==13.9.2
torch==2.4.1
transformers==4.45.2
trl==0.11.4
accelerate==0.34.2
datasets==3.0.1
tensorboard==2.18.0
openai==1.51.0
